+++ packages/react-integration/demo-app-ts/src/components/demos/EmptyStateDemo/EmptyStateDemo.tsx
@@ -11,6 +11,30 @@ import {
 import CubesIcon from '@patternfly/react-icons/dist/js/icons/cubes-icon';
 import React, { Component } from 'react';
 export class EmptyStateDemo extends Component {
+  myExtraLargeEmptyStateProps: EmptyStateProps = {
+    children: (
+      <React.Fragment>
+        <EmptyStateIcon icon={CubesIcon} />
+        <Title headingLevel="h5" size="lg">
+          Extra Large Empty State
+        </Title>
+        <EmptyStateBody>
+          This represents an the empty state pattern in Patternfly 4. Hopefully it's simple enough to use but flexible
+          enough to meet a variety of needs.
+        </EmptyStateBody>
+        <Button variant="primary">Primary Action</Button>
+        <EmptyStateSecondaryActions>
+          <Button variant="link">Multiple</Button>
+          <Button variant="link">Action Buttons</Button>
+          <Button variant="link">Can</Button>
+          <Button variant="link">Go here</Button>
+          <Button variant="link">In the secondary</Button>
+          <Button variant="link">Action area</Button>
+        </EmptyStateSecondaryActions>
+      </React.Fragment>
+    ),
+    variant: EmptyStateVariant.xl
+  };
   myLargeEmptyStateProps: EmptyStateProps = {
     children: (
       <React.Fragment>
@@ -32,7 +56,8 @@ export class EmptyStateDemo extends Component {
           <Button variant="link">Action area</Button>
         </EmptyStateSecondaryActions>
       </React.Fragment>
-    )
+    ),
+    variant: EmptyStateVariant.large
   };
   mySmallEmptyStateProps: EmptyStateProps = {
     children: (
@@ -79,8 +104,19 @@ export class EmptyStateDemo extends Component {
           <Button variant="link">Action area</Button>
         </EmptyStateSecondaryActions>
       </React.Fragment>
-    ),
-    variant: EmptyStateVariant.full
+    )
+  };
+  myFullHeightEmptyStateProps: EmptyStateProps = {
+    id: 'full-height-example',
+    children: (
+      <React.Fragment>
+        <EmptyStateIcon icon={CubesIcon} />
+        <Title headingLevel="h5" size="lg">
+          Full height empty state
+        </Title>
+        <EmptyStateBody>This represents a full height empty state pattern in Patternfly 4</EmptyStateBody>
+      </React.Fragment>
+    )
   };
 
   componentDidMount() {
@@ -90,9 +126,12 @@ export class EmptyStateDemo extends Component {
   render() {
     return (
       <React.Fragment>
-        <EmptyState>{this.myLargeEmptyStateProps.children}</EmptyState>
+        <EmptyState variant={this.myLargeEmptyStateProps.variant}>{this.myLargeEmptyStateProps.children}</EmptyState>
         <EmptyState variant={this.mySmallEmptyStateProps.variant}>{this.mySmallEmptyStateProps.children}</EmptyState>
-        <EmptyState variant={this.myFullEmptyStateProps.variant}>{this.myFullEmptyStateProps.children}</EmptyState>
+        <EmptyState>{this.myFullEmptyStateProps.children}</EmptyState>
+        <EmptyState isFullHeight id={this.myFullHeightEmptyStateProps.id}>
+          {this.myFullHeightEmptyStateProps.children}
+        </EmptyState>
       </React.Fragment>
     );
   }